Conflicts:
	src/sys/tabletUI/tabletUI.cmake

+++ src/OpenCOVER/OpenVRUI/sginterface/vruiCollabInterface.h
@@ -13,7 +13,6 @@
 #include <list>
 #include <string>
 
-//class OPENVRUIEXPORT vruiCollabInterface;
 namespace vrui
 {
 class vruiCollabInterface;
@@ -79,7 +78,7 @@ protected:
     virtual void releaseRemoteLock(const char *message);
 
 private:
-    inline int composeMessage(const char messageType, const char *message);
+    inline size_t composeMessage(const char messageType, const char *message);
 
     int interfaceType; ///< type of this interface
     std::string name; ///< symbolic name of this interface
